Continuing work on parser

+++ src/SAMPi.pl
@@ -7,11 +7,11 @@
 # puts it into CSV format and stores it in preparation for upload via SFTP
 #
 # This software works in conjunction with the SAMPiD daemon to
-# handle uploading csv data files, removal of data older than 
-# a configurable threshold and handle other miscellaneous tasks
+# handle uploading CSV data files, removal of data older than 
+# a configurable threshold and other miscellaneous tasks
 #
 # SAMPi.pl can be automatically updated by uploading a newer version
-# accessible using a configurable URL
+# to a configurable update URL
 
 # Modern Perl #
 
@@ -25,6 +25,7 @@ use Readonly; # Allows read-only constants
 use Carp; # Provides alternative warn and die functions
 
 use Sys::RunAlone; # This module ensures only one instance of this software runs concurrently
+use Sys::Hostname; # Acquire hostname 
 use Device::SerialPort; # Serial IO library 
 use LWP::Simple qw(getstore is_error); # Used to download updates
 use Digest::SHA1 qw(sha1_base64); # SHA1 checksum library
@@ -46,22 +47,28 @@ Readonly my $DIRECTORY_SEPARATOR => ($^O=~/Win/) ? "\\" : "/"; # Ternary operato
 Readonly my $CURRENT_VERSION_PATH => abs_path($0);
 Readonly my $LATEST_VERSION_PATH => File::Spec->tmpdir() . $DIRECTORY_SEPARATOR . "SAMPi.pl";
 
+# Constants used to identify which part of the data we are currently parsing
+Readonly my $UNKNOWN => -1;
+Readonly my $HEADER => 0;
+Readonly my $TRANSACTION => 1;
+Readonly my $REPORT => 2;
+Readonly my $Z1_REPORT => 3;
+Readonly my $Z2_REPORT => 4;
+Readonly my $Z3_REPORT => 5;
+Readonly my $X_REPORT => 6;
+
 my $logOpen = 0; # Flag to determine if log file has been opened
 my $logFile; # Log file descriptor, used in multiple functions
-my $serialPort; # Serial port file descriptor, as above
+my $csvFile; # CSV file descriptor, used for output
+my $serialPort; # Serial port file descriptor, used for input
+
 my $updateAvailable = 0; # Update flag
 my $postUpdateExecuted; # Filename referred to in postUpdate()
 
+my $currentDate; # Stores the current date in a string 
+my $currentEvent = $UNKNOWN; # Tracks the current type of data being parsed
+my $currentEventTime = 0;
 my $transactionCount = 0; # Counter for number of transactions per day
-
-# Keep track of which part of the data we are parsing (header, transaction, z-report or x-report)
-my $inHeader = FALSE;
-my $inTransaction = FALSE;
-my $inReport = FALSE;
-my $inZ1Report = FALSE;
-my $inZ2Report = FALSE;
-my $inZ3Report = FALSE;
-my $inXReport = FALSE;
 
 # Functions #
 
@@ -79,9 +86,12 @@ sub logMsg
         if ($logOpen == 0)
         {
             # Create a new log file each month with the month and year included in the filename
+            my $currentDay = $timestamp[3];
             my $currentMonth = $timestamp[4] + 1; # Months from localtime() are zero-indexed
             my $currentYear = $timestamp[5] + 1900; # Need to add 1900 to get the current year
             my $logFileName = "SAMPi-" . $currentMonth . "-" . $currentYear . ".log"; # Construct the filename
+
+            $currentDate = $currentDay . "-" . $currentMonth . "-"  . $currentYear;
 
             # Attempt to open the log file (located in the directory of the currently running script) in append mode
             ## no critic qw(RequireBriefOpen)
@@ -165,7 +175,7 @@ sub isBusinessHours
     else
     {
         # Write the daily transaction count to the CSV file and then reset it
-        parsedLineToCSV("TRANSACTION_COUNT", $transactionCount);
+        parsedLineToCSV("TRANSACTION_COUNT", $transactionCount); # FIXME: Wrong parameters 
         $transactionCount = 0; # Reset daily transaction count
         return FALSE;
     }
@@ -194,7 +204,7 @@ sub isUpdateAvailable
     if (compare ($CURRENT_VERSION_PATH, $LATEST_VERSION_PATH) == 1) 
     {
         # So we report this availability to the caller 
-        return TRUE;
+        return FALSE;
 	}
 
     else
@@ -213,7 +223,13 @@ sub updateAndReload
         logMsg("Update found, overwriting $CURRENT_VERSION_PATH with $LATEST_VERSION_PATH");
         copy($LATEST_VERSION_PATH, $CURRENT_VERSION_PATH);
         logMsg("Restarting...");
-        unlink $postUpdateExecuted; # Remove file which signifies we have run postUpdate() before
+
+        # Remove file which signifies we have run postUpdate() before if it exists
+        if (-e $postUpdateExecuted)
+        {
+            unlink $postUpdateExecuted; 
+        }
+
         exec $0; # Exec call replaces the currently running process with the new version
     }
 
@@ -248,7 +264,7 @@ sub postUpdate
         logMsg("postUpdate() call requested, executing postUpdateCode");
 
         # String eval is not recommended but is the only way to run this code dynamically
-        ## no critic (ProhibitStringyEval)
+        ## no critic qw(ProhibitStringyEval)
         my $postUpdateValid = eval $postUpdateCode;
 
         # Detect errors and log accordingly
@@ -272,68 +288,113 @@ sub postUpdate
 # This function is responsible for the first-stage of parsing a line of serial data with regard to the current state of the data
 # This stage differentiates between transaction reports, z-reports and x-reports and separates events for further processing
 # Into CSV format in the parsedLineToCSV() function
+# TODO: Finish implementing this function
 sub parseLine
 {
     # The line passed in as a parameter will be accessible as $dataLine
     my ($dataLine) = @_;
 
-    print "Parsing $dataLine";
+    # Replace any 0x9c hex values returned by the serial port with a '£'
+    $dataLine =~ s/\x{9c}/£/gx;
+
+    print "Input is: $dataLine";
 
     # Handle headers
     # If the line begins with a date in dd/mm/yyyy format, it is a header
     if ($dataLine =~ /^[0-9][0-9]\/[0-9][0-9]\/[0-9][0-9][0-9][0-9]/x)
     {
-        $inHeader = TRUE;
+        $currentEvent = $HEADER;
 
-        # Extract the time into the $1 variable
+        # Extract the time from the line into the $1 variable
         $dataLine =~ /([0-9][0-9]:[0-9][0-9])/x;
 
         # Check that the extraction succeeded 
         if ($1)
         {
-            parsedLineToCSV($1); # Write the time of the event to the CSV file 
+            $currentEventTime = $1; # Store the time of the current event
         }
 
         return;
     }
 
     # Handle the first line after a header
-    if ($inHeader)
+    if ($currentEvent == $HEADER)
     {
-        # If line after header does not contain 'REPORT', we are reading a transaction event
-        if (index ($dataLine, 'REPORT') == -1)
+        # If the line after the header contains 'REPORT', we are reading (one of four types of) report
+        if (index($dataLine, 'REPORT') != -1)
         {
-            $inTransaction = TRUE;
-            $transactionCount++; # Increment globally accessible daily transaction counter
+            $currentEvent = $REPORT;
+            return; # We need more context to see what type of report this is so we return
         }
 
-        else
-        {
-            $inReport = TRUE;
-        }
-
-        $inHeader = FALSE;
+        # Otherwise, we are reading a transaction event
+        $currentEvent = $TRANSACTION;
+        $transactionCount++; # Increment globally accessible daily transaction counter
     }
 
     # Handle a line that is part of a transaction
-    if ($inTransaction)
+    if ($currentEvent == $TRANSACTION)
     {
-        # Match a line containing only one or more dashes (these are event separators)
-        if ($dataLine =~ /^-+$/x)
-        {
-            # This means we have reached the end of a transaction so we disable the flag and return
-            $inTransaction = FALSE;
-            return;
-        }
-
-        # If the line contains a transaction, replace the 0x9c hex value returned by the serial port with a '£'
-        $dataLine =~ s/\x{9c}/£/gx;
-
         # Separate the data into a key (identifier preceding a price) and a value (price) to use as input for the second parser stage
         my ($transactionKey, $transactionValue) = split('£', $dataLine, 2);
 
+        # Treat the TOTAL line differently to a standard transaction line
+        if (index($transactionKey, "TOTAL") != -1)
+        {
+            parseLineToCSV($transactionValue);
+            return;
+        }
+
         # Pass the separated data to the second-stage parser for formatting into CSV
         parsedLineToCSV($transactionKey, $transactionValue);
+        return;
+    }
+
+    # Handle a line that specifies the type of report being processed
+    if ($currentEvent == $REPORT)
+    {
+        ## no critic qw(ProhibitCascadingIfElse)
+        if ($dataLine =~ /^FINANCIAL/x)
+        {
+            $currentEvent = $Z1_REPORT;
+        }
+
+        elsif ($dataLine =~ /^TIME/x)
+        {
+            $currentEvent = $Z2_REPORT;
+        }
+
+        elsif ($dataLine =~ /^ALL PLU/x)
+        {
+            $currentEvent = $Z3_REPORT;
+        }
+
+        elsif ($dataLine =~ /^X 1/x)
+        {
+            $currentEvent = $X_REPORT;
+        }
+
+        return;
+    }
+
+    if ($currentEvent == $Z1_REPORT)
+    {
+        print("Z1 Reports not yet implemented\n");
+    }
+
+    if ($currentEvent == $Z2_REPORT)
+    {
+        print("Z2 Reports not yet implemented\n");
+    }
+
+    if ($currentEvent == $Z3_REPORT)
+    {
+        print("Z3 Reports not yet implemented\n");
+    }
+
+    if ($currentEvent == $X_REPORT)
+    {
+        print("X Reports not yet implemented\n");
     }
 
     return;
@@ -342,16 +403,59 @@ sub parseLine
 # This function is the second-stage parser and accepts processed data, performs further parsing
 # and converts the data into CSV format in preparation for uploading 
 # TODO: Finish implementing this function
+## no critic qw(RequireArgUnpacking)
 sub parsedLineToCSV
 {
-    my $numParameters = @_; 
+    # Create an appropriately named CSV file and open it in append mode if it does not already exist
+    if (!$csvFile)
+    {
+        initiliaseCSVFile();
+    }
+
+    my $numParameters = @_;
+
+    # Handle single values
+    if ($numParameters == 1)
+    {
+        my ($transactionKey) = @_;
+
+        if (index($transactionKey, 'TOTAL') != -1)
+        {
+            print $csvFile "Total is $transactionKey\n";
+        }
+    }
 
     # Handle key value pairs 
     if ($numParameters == 2)
     {
+        my ($transactionKey, $transactionValue) = @_;
 
+        print $csvFile "$transactionKey,£$transactionValue,";
     }
     
+    return;
+}
+
+# This function creates a CSV file in the local ecr_data directory with a list of predefined headings
+# and named with the date of creation and hostname of the machine SAMPi is running on
+sub initiliaseCSVFile
+{
+    my $hostname = hostname();
+    my $csvFilename = $hostname . $currentDate;
+
+    ## no critic qw(RequireBriefOpen)
+    open($csvFile, '>>', $csvFilename) or die "Error opening CSV file: $!";
+
+    my @csvHeadings = ("Heading 1", "Heading 2", "Heading 3");
+
+    # Write the headings to the file
+    foreach (@csvHeadings)
+    {
+        print $csvFile "$_,";
+    }
+
+    print $csvFile "\n";
+
     return;
 }
 
@@ -374,7 +478,7 @@ sub processData
         # Read ECR data over serial and process it during business hours
         while ($storeIsOpen)
         {
-            # Set parameters for serial I/O, look for EOL characters so we read in line by line
+            # Set parameters for serial IO, look for EOL characters so we read in line by line
             $serialPort->are_match( "\r", "\n" );
          
             # Wait until we have read a line of data 
@@ -419,7 +523,7 @@ sub processData
                     last;
                 }
 
-                sleep($UPDATE_CHECK_DELAY_MINUTES * 60);
+                sleep($UPDATE_CHECK_DELAY_MINUTES * 60); # The parameter must be in seconds
             }
         }
     }
